+++ x11/x11.cu
@@ -60,7 +60,7 @@ extern void x11_echo512_cpu_init(int thr_id, uint32_t threads);
 
 extern void quark_compactTest_cpu_init(int thr_id, uint32_t threads);
 extern void quark_compactTest_cpu_hash_64(int thr_id, uint32_t threads, uint32_t startNounce, uint32_t *inpHashes,
-                                          uint32_t *d_noncesTrue, size_t *nrmTrue, uint32_t *d_noncesFalse, size_t *nrmFalse, int order);
+                                          uint32_t *d_noncesTrue, uint32_t *nrmTrue, uint32_t *d_noncesFalse, uint32_t *nrmFalse, int order);
 
 // X11 CPU Hash
 extern "C" void x11hash(void *output, const void *input)
