+++ x13/x13.cu
@@ -70,7 +70,7 @@ extern void  x11_simd512_cpu_free(int32_t thr_id);
 
 extern void quark_compactTest_cpu_init(int thr_id, uint32_t threads);
 extern void quark_compactTest_cpu_hash_64(int thr_id, uint32_t threads, uint32_t startNounce, uint32_t *inpHashes, 
-                                          uint32_t *d_noncesTrue, size_t *nrmTrue, uint32_t *d_noncesFalse, size_t *nrmFalse, int order);
+                                          uint32_t *d_noncesTrue, uint32_t *nrmTrue, uint32_t *d_noncesFalse, uint32_t *nrmFalse, int order);
 
 // X13 Hashfunktion
 extern "C" void x13hash(void *output, const void *input)
